Activator: include instead of auto include.
No need to go around and stretch the code to not have methods available
if we haven't activated this for a model. As this will most likely only
be used for one or two models in a project, it makes more sense to have
this being included.

Instead of magically having `act_as_activator` - which is a horrible
name anyways - do all the work, we now use `include Activator` to get
our methods that we want.